// packages/data/src/static/requirements/data/gdpr.types.ts
export type gdprRequirementIds =
	| "A4" // Data protection by design and by default
	| "A5" // Principles relating to processing of personal data
	| "A6" // Lawfulness of processing
	| "A7" // Conditions for consent
	| "A12" // Transparent information, communication and modalities for the exercise of the rights of the data subject
	| "A13" // Information to be provided where personal data are collected from the data subject
	| "A14" // Information to be provided where personal data have not been obtained from the data subject
	| "A15" // Right of access by the data subject
	| "A16" // Right to rectification
	| "A17" // Right to erasure (‘right to be forgotten’)
	| "A18" // Right to restriction of processing
	| "A19" // Right to notification of erasure
	| "A20" // Right to data portability
	| "A21" // Right to object
	| "A25" // Data protection by design and by default
	| "A30" // Records of processing activities
	| "A32" // Security of processing
	| "A33" // Notification of a personal data breach to the supervisory authority
	| "A34" // Communication of a personal data breach to the data subject
	| "A35"; // Data protection impact assessment